llvm.cfg command refactoring
Now it lives in a separate file
